sig: Deprecate IplImage.h header and move the actual file in impl
