+++ src/libYARP_sig/src/ImageCopy.cpp
@@ -9,7 +9,7 @@
 
 #include <yarp/os/Log.h>
 #include <yarp/sig/Image.h>
-#include <yarp/sig/IplImage.h>
+#include <yarp/sig/impl/IplImage.h>
 
 #include <cstring>
 #include <cstdio>
